requirements.txt
Python == 3.7
tensorflow-gpu == 2.1.0
scikit-learn >= 0.23.1
scipy >= 1.5.0
numpy >= 1.18.5
pydot >= 1.4.1
hdf5 >= 1.10.4
h5py >= 2.10.0
matplotlib >= 3.2.2
graphviz >= 2.38
mne >= 0.20.7
braindecode == 0.2